Describe the manual unpublishing process
We no longer have support for 'unpublishing' measures built-in to the
Publisher. We made this choice explicitly so that it is not a step that
can be easily done - once we publish a page, it's out there in the
public eye and to retract it could put us in a bad light. We should
publish an updated version that corrects and mistakes, instead.

However, we anticipate there *might* be some cases where we need to
unpublish a thing. Where this is the case, it will probably be a
high-stress situation, so documenting the manual process for
unpublishing seems key.